Se separan los eventos a sus propios archivos
