+++ src/blink.c
@@ -1,4 +1,5 @@
 #include "my_rtos.h"
+#include "my_rtos_events.h"
 #include "sapi.h"
 #include "blink.h"
