+++ src/my_rtos.c
@@ -10,15 +10,11 @@ static bool MyRtos_initTask(taskControl_t *task, taskID_t ID);
 
 static void MyRtos_idleTask(void *param);
 
-static void MyRtos_schedulerUpdate(void);
-
 static void MyRtos_delaysUpdate(void);
 
 static void MyRtos_returnHook(void);
 
 static bool MyRtos_getReadyTask(taskID_t *ID);
-
-static void MyRtos_addReadyTask(taskID_t ID);
 
 static void MyRtos_updateReadyList(taskID_t ID);
 
@@ -90,13 +86,41 @@ void MyRtos_StartOS(void) {
 }
 
 
+/**
+ * @brief Returns the current task's ID
+ * 
+ * @return taskID_t Task's ID
+ */
+taskID_t MyRtos_GetCurrentTask(void) {
+   return currentTask;
+}
+
+
+/**
+ * @brief Calls the scheduler for a context switching
+ * 
+ */
+void MyRtos_SchedulerUpdate(void) {
+   // Activate PendSV for context switching
+   SCB->ICSR = SCB_ICSR_PENDSVSET_Msk;
+
+   // Instruction Synchronization Barrier: this way we make sure that all
+   // pipelined instructions are executed
+   __ISB();
+      
+   // Data Synchronization Barrier: this way we make sure that all
+   // memory accesses are completed
+   __DSB();
+}
+
+
 void MyRtos_DelayMs(uint32_t ms) {
    if (currentTask != MY_RTOS_TASK_NONE && ms != 0 &&
        MyRtos_TasksList[currentTask].state == TASK_RUNNING)
    {
       MyRtos_TasksList[currentTask].delay = ms;
       MyRtos_TasksList[currentTask].state = TASK_BLOCKED;
-      MyRtos_schedulerUpdate();
+      MyRtos_SchedulerUpdate();
    }
 }
 
@@ -139,88 +163,6 @@ uint32_t MyRtos_GetNextContext(uint32_t currentSP) {
    return MyRtos_TasksList[currentTask].stackPointer;
 }
 
-
-/**
- * @brief Initializes an event for it's future use.
- * 
- * @param event Pointer to the event to be initialized
- * @return true Event was initialized
- * @return false Problem initializing event
- */
-bool MyRtos_EventInit(event_t *event) {
-   static eventID_t ID = 0;
-
-   // Initialize event structure
-   event->state = EVENT_INITIALIZED;
-   event->eventID = ID;
-   event->taskID = MY_RTOS_TASK_NONE;
-
-   // Increment event ID for next one
-   ID++;
-
-   return true;
-}
-
-
-/**
- * @brief 
- * 
- * @param event Waits for an event to be sent. The task that calls this function
- *              will be blocked until the event is sent.
- * @return true Event waited successfully
- * @return false Could not wait for event (check event state)
- */
-bool MyRtos_EventWait(event_t *event) {
-   
-   // Check if state can be used
-   if (event->state != EVENT_INITIALIZED) {
-      return false;
-   }
-
-   // Save current task ID into the event to unblock it
-   event->taskID = currentTask;
-
-   // Mark event as pending
-   event->state = EVENT_PENDING;
-
-   // Mark current task as blocked
-   MyRtos_TasksList[currentTask].state = TASK_BLOCKED;
-
-   // Call the scheduler for context switching
-   MyRtos_schedulerUpdate();
-   
-   return true;
-}
-
-
-/**
- * @brief Sends an event to unblock a task that was waiting for it.
- * 
- * @param event Pointer to the event to be sent
- * @return true Event was sent successfully
- * @return false Event could not be sent
- */
-bool MyRtos_EventSend(event_t *event) {
-
-   // Check if state was pending
-   if (event->state != EVENT_PENDING) {
-      return false;
-   }
-
-   // Add the blocked task as ready
-   MyRtos_addReadyTask(event->taskID);
-
-   // Reset the event's task id
-   event->taskID = MY_RTOS_TASK_NONE;
-
-   // Reset the event's state
-   event->state = EVENT_INITIALIZED;
-
-   // Call the scheduler for context switching
-   MyRtos_schedulerUpdate();
-
-   return true;
-}
 ///////////////////////////////////////////////////////////////////////////////
 // Internal functions definitions
 ///////////////////////////////////////////////////////////////////////////////
@@ -255,7 +197,7 @@ static bool MyRtos_getReadyTask(taskID_t *ID) {
  * 
  * @param ID The ID of the task to be added to the list
  */
-static void MyRtos_addReadyTask(taskID_t ID) {
+void MyRtos_AddReadyTask(taskID_t ID) {
    uint32_t slot = 0;
    taskPriority_t priority = MyRtos_TasksList[ID].basePriority;
 
@@ -327,7 +269,7 @@ static void MyRtos_updateReadyList(taskID_t ID) {
 static bool MyRtos_initTask(taskControl_t *task, taskID_t ID) {
    MyRtos_initStack(task);
 
-   MyRtos_addReadyTask(ID);
+   MyRtos_AddReadyTask(ID);
 
    return true;
 }
@@ -374,24 +316,11 @@ static void MyRtos_delaysUpdate(void) {
       if (MyRtos_TasksList[i].delay != 0) {
          MyRtos_TasksList[i].delay--;
          if (MyRtos_TasksList[i].delay == 0) {
-            MyRtos_addReadyTask(i);
+            MyRtos_AddReadyTask(i);
          }
       }
       i++;
    }
-}
-
-static void MyRtos_schedulerUpdate(void) {
-   // Activate PendSV for context switching
-   SCB->ICSR = SCB_ICSR_PENDSVSET_Msk;
-
-   // Instruction Synchronization Barrier: this way we make sure that all
-   // pipelined instructions are executed
-   __ISB();
-      
-   // Data Synchronization Barrier: this way we make sure that all
-   // memory accesses are completed
-   __DSB();
 }
 
 void SysTick_Handler(void) {
@@ -400,5 +329,5 @@ void SysTick_Handler(void) {
 
    MyRtos_delaysUpdate();
 
-   MyRtos_schedulerUpdate();
+   MyRtos_SchedulerUpdate();
 }
